create base node "lib" module
