+++ applications/tari_base_node/src/cli.rs
@@ -28,7 +28,7 @@ use tari_common::configuration::{ConfigOverrideProvider, Network};
 #[clap(author, version, about, long_about = None)]
 #[clap(propagate_version = true)]
 #[allow(clippy::struct_excessive_bools)]
-pub(crate) struct Cli {
+pub struct Cli {
     #[clap(flatten)]
     pub common: CommonCliArgs,
     /// Create a default configuration file if it doesn't exist
