+++ applications/tari_base_node/src/recovery.rs
@@ -55,7 +55,7 @@ use tari_core::{
     },
 };
 
-use crate::config::{BaseNodeConfig, DatabaseType};
+use crate::{BaseNodeConfig, DatabaseType};
 
 pub const LOG_TARGET: &str = "base_node::app";
